Enable sorting RSS feed by tested_at date
By default RSS feed is sorted by created_at date. However, every now and
then results are uploaded later and they may not be of particular
interest to anyone, but still appear as the first element in the RSS
feed. By enabling sorting by tested_at date users can see a feed that is
in the same order as report listing in the UI, ie. showing the newest
results first.