Added configurable links to upper header
The three links in the #upper_header section are not necessarily needed
by all, or they need to point to some organization specific target. Thus
added configurable targets for the links.

The link can be a mailto-link (a nice envelope shown as icon), or any
other piece of string (the external icon shown). If left empty, the
whole item will be removed.