Implemented test case lists for cumulative reports.
Each feature section now has a list of test cases, which contains the testcase's name and final
accumulated result. Since the number of test cases can be quite large, test cases are formatted to
json as a list instead of an object to avoid unnecessary key duplication. So each test case is
(currently) a list of three elements:

[name, tc_id, result]

(TC_ID is currently always 'null' in the current implementation)

When new fields are added to the test case list, they should always be added to the end of the so
that indinces are preserved for existing users of the API.