Added googletest XML support to XML parser
Added (possibly preliminayr) Google Test Framework XML output support to
XMLResultFileParser. The input data comes from documentation but it
likely lacks some features that exist.

The current implementation is anyways such that if the test case status
is 'run' and it has no <failure> nodes as children the case is passed.
If it has <failure> nodes it is failed, and if the status is something
else than run the result will be N/A.