Fix test case result editing
Test case result editing was a bit broken after having added support to
custom results and changed the form to edit the result name. The change
was done but UI did not reflect it. Now the previous functionality is
restored, ie. the result changes and gets correct class right away.

Also fixed removed and recovered test case result editing. The old
change handler was not removed in unlinkTestCaseButtons which caused the
result editing not working after having removed and restored a case.