Use routes as exception_app
To enable rendering error pages with custom stylesheets we need to catch
routing exceptions. As of Rails 3.0 this changed a lot and currently it
seems no really good method exists. Of all the ones in the internet this
seemed to work quite well and did not interfere with generated routes:

http://techoctave.com/c7/posts/36-rails-3-0-rescue-from-routing-error-solution