Added trailing wildcard option to RSS feed
Need to be able to combine results to single feed, e.g. if user wants
to see test results for test set Acceptance from both 1.1 and 1.2
releases he can now ask for /1.*/*/Acceptance/rss to get that. The
wildcard for target does not actually need to be an asterisk because
the filter code works so that it tries to find by name, and if it does
not find anything the target value will be dropped from the filter
(i.e. it just happens to work the way needed)