Set the feed uri from first article if wildcards are used
If using wildcards the selected release and profile do not match the
feed (possibly at all). Also, it is pretty much impossible to set
correct feed title in the current format when articles in fact come from
multiple releases/targets. Thus, just use the values from the first
article and be done with it.