Manually expire cache and index
If a test set contains cases with measurements, MeegoTestCase.save will
be called for each such case and thus caches would be cleared multiple
times. This is just a small improvement, the major issue is the very
poor saving performance when you have e.g. 1000 cases, each with a
single measurement.