Hide report summary if defined so
If hide_summary is set to true the report summary section is not shown.
This was requested by a user organization. It was not just hidden using
css because in most cases the summary should be shown. This feature is
sort of prequisite for another feature request of enabling metrics
reporting on test report level (similar to measurements but on higher
level).