xUnit test case
It seems, although it's not certain, that xUnit test result xml contains
testcases, and they may or may not have failure or error child nodes.
Added a test with such file.

It seems, although it's not certain, that this differs from Google Test
FW XML results so that Google testcases have result attribute whereas
xUnit ones do not.