Added support for xUnit test XML
Now a testcase without a result or status is considered to be an xUnit
test case, and those are considered passed unless they have failures or
errors as children.

This has an effect of accepting MeeGo test cases missing result, but
currently that does not seem to be too big of a tradeoff.