Changed order of head elements due to IE9
It seems that if title is the first element in head, and X-UA-... meta
tag is included directly after it, IE9 will go to IE9 document mode also
when running locally without nginx setting the IE=edge header.

Documentation to be updated still to include the header, but seems to be
working now with both IE8 and IE9.