Use new external service config file
Changed to initializer to handle new config file and, if missing, the
old bugzilla.yml config file.

In addition renamed the constant and took it in use in existing code.
Currently the first element is used in place of the old BUGZILLA_CONFIG,
i.e. defining multiple services has no effect on the functionality.