Added new field "Patches Included"
Added a field with purpose to list patches included in the software
build being tested. It is quite common use case for organizations to
track what they are testing by reporting the patch IDs, and now that we
can link to random services such as Gerrit this feature is quite
meaningful.