Run external service calls in parallel
When requesting data from multiple external services each http call
blocked the execution and thus caused the response be slowish. Now the
external service handlers are classes including Celluloid, and the
controller collects a list of futures. This way all the handlers are
launched in parallel which will speed up the response.