Checks for some invalid data
Check release and profile are found when creating the JSON report.
Naturally there should not be such reports but real life has shown that
if report grouping is changed not all reports not accessible anymore are
necessarily moved/removed by the admin.

So instead of failing miserably just skip such reports.