Don't replace HTML when editing test cases
Previously editing test case comments received HTML fragment from
backend and placed that in the table cell. This fragment did not however
have the edit mode set and thus re-editing a comment did not work
properly (input was empty).

Now just return HTTP 200 OK from the backend and remove the form element
in the frontend so the markup formatted comments is not lost.