Don't try to upload to QA Dashboard if not enabled
Previously cap deploy:setup did ask if user wanted to set up exporting
to QA Dashboard but didn't really follow the answer - upload was always
tried but since the token/url were incorrect it just didn't work. This
makes no sense so now we check for enabled flag in the file and just
return true if it has been set and is set to false.