No 404s from ReportGroupController
Based on user feedback to most common scenario for receiving a 404 from
QA Reports is that you're switching to another release version while
currently being on the report group listing page.

If the other release version does not have reports for that particular
target/testset/product combination user was greeted with a 404.

This commit changes that so that first we try to broaden the scope first
(try with target/testset/product, then with target/testset, then with
just target), and finally if we still don't get reports to show we will
redirect to root page which will then show the index page for the
selected release version.

This does introduce a minor oddity that if you have target X for release
version Y but there are no reports, clicking the target link in the
index page will redirect back to the index page.

Fixes #19