Show a no reports notification
If using hide_empty_targets and there are no reports to show for a
release the page looks confusing. We now show a notification about this.

Also removed the tab clicking stuff from initTabs because it caused the
json being downloaded three times (when coming to a page with hash
fragment with all as scope), and also it caused the page to be in wrong
state (selected scope was all but the last JSON download was done with
scope recent). Spine route setup executes routes and we now set the
attributes there.