Test data for grouped serial measurements
The original test result XML schema has attribute "group" for a series
element. It has not been used in QA Reports, and the original purpose of
the attribute is unknown (no documentation in the schema or in the
commit message, and since bugs.meego.com is no longer online the
original item cannot be checked).

So a brave choise is made to use this attribute for grouping serial
measurements together so that we can draw a graph showing more than one
data series in single graph.