Support for grouping series with non-matching measurements
Serial measurement groups can now contain series with measurements that
are not using the same x axis.

If the series are interval based all representations begin from the
beginning of "time", and some end earlier.

If the series are timestamp based an x axis that can hold all values is
created and the data series on a chart will begin at some point of time
and end at another.